remove unecessary interface
